:recycle: Update front-end travel auth -> pre-approval profile code.
