+++ web/src/modules/travel-authorizations/components/edit-my-travel-authorization-details-page/ApprovalsFormCard.vue
@@ -55,8 +55,8 @@
             cols="12"
             md="4"
           >
-            <TravelAuthorizationPreApprovalSelect
-              v-model="travelAuthorization.preApprovalId"
+            <TravelAuthorizationPreApprovalProfileSelect
+              v-model="travelAuthorization.preApprovalProfileId"
               :department="department"
               dense
               outlined
@@ -103,7 +103,7 @@
 import { mapActions, mapGetters } from "vuex"
 
 import SearchableUserEmailCombobox from "@/components/SearchableUserEmailCombobox"
-import TravelAuthorizationPreApprovalSelect from "@/components/travel-authorization-pre-approvals/TravelAuthorizationPreApprovalSelect"
+import TravelAuthorizationPreApprovalProfileSelect from "@/components/travel-authorization-pre-approval-profiles/TravelAuthorizationPreApprovalProfileSelect"
 import EstimatedCostTextField from "@/modules/travel-authorizations/components/EstimatedCostTextField"
 import TravelAuthorizationActionLogsTable from "@/modules/travel-authorizations/components/TravelAuthorizationActionLogsTable"
 
@@ -118,7 +118,7 @@ export default {
     SearchableUserEmailCombobox,
     SubmitToSupervisorButton,
     TravelAuthorizationActionLogsTable,
-    TravelAuthorizationPreApprovalSelect,
+    TravelAuthorizationPreApprovalProfileSelect,
   },
   props: {
     travelAuthorizationId: {
