+++ web/src/modules/travel-authorizations/components/edit-travel-authorization-details-page/ApprovalsFormCard.vue
@@ -55,8 +55,8 @@
             cols="12"
             md="4"
           >
-            <TravelAuthorizationPreApprovalSelect
-              v-model="travelAuthorization.preApprovalId"
+            <TravelAuthorizationPreApprovalProfileSelect
+              v-model="travelAuthorization.preApprovalProfileId"
               :department="department"
               dense
               outlined
@@ -108,6 +108,7 @@ import useExpenses, { TYPES as EXPENSE_TYPES } from "@/use/use-expenses"
 import useUser from "@/use/use-user"
 
 import SearchableUserEmailCombobox from "@/components/SearchableUserEmailCombobox"
+import TravelAuthorizationPreApprovalProfileSelect from "@/components/travel-authorization-pre-approval-profiles/TravelAuthorizationPreApprovalProfileSelect"
 import EstimatedCostTextField from "@/modules/travel-authorizations/components/EstimatedCostTextField"
 import TravelAuthorizationActionLogsTable from "@/modules/travel-authorizations/components/TravelAuthorizationActionLogsTable"
